+++ lib/src/televerse/context/callback_query.dart
@@ -1,13 +1,13 @@
 part of televerse.context;
 
-class CallbackQueryContext extends Context {
+class CallbackQueryContext extends Context with CallbackQueryMixin {
   /// The incoming callback query.
   CallbackQuery query;
 
   CallbackQueryContext(
     Televerse t,
     this.query, {
-    Update? update,
+    required Update update,
   }) : super(
           t,
           update: update,
@@ -24,22 +24,6 @@ class CallbackQueryContext extends Context {
 
   /// The user of the query.
   User? get from => message?.from;
-
-  /// Answer the callback query.
-  Future<void> answer(
-    String? text, {
-    bool showAlert = false,
-    String? url,
-    int cacheTime = 0,
-  }) async {
-    await _televerse.answerCallbackQuery(
-      query.id,
-      text: text,
-      showAlert: showAlert,
-      url: url,
-      cacheTime: cacheTime,
-    );
-  }
 
   List<RegExpMatch>? matches;
 }