+++ lib/televerse.dart
@@ -21,6 +21,10 @@ part 'src/utils/date.dart';
 part 'src/utils/http.dart';
 part 'src/utils/utils.dart';
 part 'src/televerse/filters/filters.dart';
+part 'src/televerse/context/mixins/message_mixin.dart';
+part 'src/televerse/context/mixins/management.dart';
+part 'src/televerse/context/mixins/callback_mixin.dart';
+part 'src/televerse/context/mixins/inline_mixin.dart';
 
 /// The main class of the library.
 ///